Implement Get Template button (BL-9584)
A variant of TranslateButton for templates.
The common functionality is moved to DownloadToBloomButton.txt
